test for gallery

+++ test/manual-docker/build.ps1
@@ -1,6 +1,21 @@
 
-$workingDir = $PSScriptRoot
-Remove-Item -Path "$workingDir\LabViewHelpers" -Recurse -Force -ErrorAction SilentlyContinue
-Copy-Item -Path "$workingDir\..\..\LabViewHelpers" -Destination "$workingDir\LabViewHelpers" -Recurse -Force
+param(
+    [switch]$gallery
+)
 
-docker build -t labview-dev-image -f $workingDir\Dockerfile $workingDir
+$workingDir = $PSScriptRoot
+
+if ($gallery) {
+    Write-Host "Building with LabViewHelpers from PowerShell Gallery..."
+    $dockerFilePath = "$PSScriptRoot\DockerfileFromGallery"
+    docker build -t labview-dev-image -f $dockerFilePath $workingDir
+} else {
+    Write-Host "Building with LabViewHelpers from local source..."
+
+    $dockerFilePath = "$PSScriptRoot\DockerfileFromSource"
+
+    Remove-Item -Path "$workingDir\LabViewHelpers" -Recurse -Force -ErrorAction SilentlyContinue
+    Copy-Item -Path "$workingDir\..\..\LabViewHelpers" -Destination "$workingDir\LabViewHelpers" -Recurse -Force
+
+    docker build -t labview-dev-image -f $dockerFilePath $workingDir
+}
